Extract ServerInDocker
